Moved general handlers from slash handler
