+++ src/main/java/com/honiism/discord/lemi/commands/handler/UserCategory.java
@@ -17,15 +17,11 @@
  * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.commands.slash.handler;
+package com.honiism.discord.lemi.commands.handler;
 
-public enum CommandCategory {
-    MAIN, // no slash category
-    CURRENCY,
-    FUN,
-    GAMES,
-
-    MODS,
+public enum UserCategory {
+    DEV,
     ADMINS,
-    DEV
+    MODS,
+    USERS
 }
