+++ src/main/java/com/honiism/discord/lemi/commands/slash/currency/Bankrob.java
@@ -19,13 +19,13 @@
 
 package com.honiism.discord.lemi.commands.slash.currency;
 
+import com.honiism.discord.lemi.commands.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Random;
 
-import com.honiism.discord.lemi.commands.slash.handler.CommandCategory;
-import com.honiism.discord.lemi.commands.slash.handler.UserCategory;
 import com.honiism.discord.lemi.utils.currency.CurrencyTools;
 import com.honiism.discord.lemi.utils.currency.WeightedRandom;
 import com.honiism.discord.lemi.utils.misc.EmbedUtils;
