+++ src/main/java/com/honiism/discord/lemi/commands/slash/currency/Inventory.java
@@ -19,14 +19,14 @@
 
 package com.honiism.discord.lemi.commands.slash.currency;
 
+import com.honiism.discord.lemi.commands.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.concurrent.TimeUnit;
 
 import com.honiism.discord.lemi.Lemi;
-import com.honiism.discord.lemi.commands.slash.handler.CommandCategory;
-import com.honiism.discord.lemi.commands.slash.handler.UserCategory;
 import com.honiism.discord.lemi.utils.currency.CurrencyTools;
 import com.honiism.discord.lemi.utils.misc.EmbedUtils;
 import com.honiism.discord.lemi.utils.misc.Tools;
