+++ src/main/java/com/honiism/discord/lemi/commands/slash/handler/ISlashCmd.java
@@ -21,6 +21,9 @@ package com.honiism.discord.lemi.commands.slash.handler;
 
 import java.util.List;
 
+import com.honiism.discord.lemi.commands.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.UserCategory;
+
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.MessageEmbed;
