+++ src/main/java/com/honiism/discord/lemi/commands/slash/handler/SlashCmd.java
@@ -22,6 +22,8 @@ package com.honiism.discord.lemi.commands.slash.handler;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.honiism.discord.lemi.commands.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.utils.misc.CustomEmojis;
 import com.honiism.discord.lemi.utils.misc.Tools;
