+++ src/main/java/com/honiism/discord/lemi/commands/slash/staff/admins/Embed.java
@@ -22,9 +22,9 @@ package com.honiism.discord.lemi.commands.slash.staff.admins;
 import java.util.Arrays;
 import java.util.HashMap;
 
-import com.honiism.discord.lemi.commands.slash.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
-import com.honiism.discord.lemi.commands.slash.handler.UserCategory;
 import com.honiism.discord.lemi.database.managers.LemiDbEmbedManager;
 import com.honiism.discord.lemi.utils.customEmbeds.EmbedTools;
 import com.honiism.discord.lemi.utils.misc.Tools;
