+++ src/main/java/com/honiism/discord/lemi/commands/slash/staff/dev/Shutdown.java
@@ -23,9 +23,9 @@ import java.util.Arrays;
 
 import com.honiism.discord.lemi.Config;
 import com.honiism.discord.lemi.Lemi;
-import com.honiism.discord.lemi.commands.slash.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
-import com.honiism.discord.lemi.commands.slash.handler.UserCategory;
 import com.honiism.discord.lemi.listeners.BaseListener;
 
 import org.slf4j.Logger;
