+++ src/main/java/com/honiism/discord/lemi/commands/slash/staff/mods/ModifyBal.java
@@ -22,9 +22,9 @@ package com.honiism.discord.lemi.commands.slash.staff.mods;
 import java.util.Arrays;
 import java.util.HashMap;
 
-import com.honiism.discord.lemi.commands.slash.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.CommandCategory;
+import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
-import com.honiism.discord.lemi.commands.slash.handler.UserCategory;
 import com.honiism.discord.lemi.utils.currency.CurrencyTools;
 import com.honiism.discord.lemi.utils.misc.Tools;
